+++ circuitforge_core/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.12.0"
+__version__ = "0.13.0"
 
 try:
     from circuitforge_core.community import CommunityDB, CommunityPost, SharedStore